Refactor: remove old welcome message code
