+++ main.js
@@ -113,10 +113,7 @@ export function bootstrap (config, cb) {
 				}
 
 				if (config.features.welcomeMessage && !isInitialized('welcomeMessage')) {
-					let version = flags.get('newFooter') ? 'new' : 'old';
-					flags.get('welcomePanel') && welcomeMessage[version].init({
-						enableOverlay: flags.get('myFTOnboardingOverlay')
-					});
+					flags.get('welcomePanel') && welcomeMessage.init();
 					initializedFeatures.welcomeMessage = true
 				}
