+++ welcome-message/test/useragents.spec.js
@@ -1,4 +1,4 @@
-const useragent = require('../new/useragent');
+const useragent = require('../useragent');
 
 const webAppIncompatibleIos = [
 	'Mozilla/5.0 (iPhone; CPU iPhone OS 6_1_4 like Mac OS X) AppleWebKit/536.26 (KHTML, like Gecko) Version/6.0 Mobile/10B350 Safari/8536.25',
